Do not hard code flavors
This was originally done because I noticed an error message referencing
ironic and I thought we may have needed to specify a custom resource.
The predefined flavors allocate generous amounts of DISK,CPU and RAM
which leads to the need to reduce concurrency. Using more constrained
flavors will allow us to increase the concurrency.